+++ google_sheets/app.py
@@ -11,6 +11,7 @@ from fastapi.responses import RedirectResponse
 from googleapiclient.errors import HttpError
 
 from . import __version__
+from .data_processing import process_data_f, validate_data
 from .db_helpers import get_db_connection
 from .google_api import (
     build_service,
@@ -290,43 +291,6 @@ MANDATORY_KEYWORD_TEMPLATE_COLUMNS = [
 ]
 
 
-def validate_data(df: pd.DataFrame, mandatory_columns: List[str], name: str) -> str:
-    error_msg = ""
-    if len(df.columns) != len(set(df.columns)):
-        error_msg = f"""Duplicate columns found in the {name} data.
-Please provide unique column names.
-"""
-    if not all(col in df.columns for col in mandatory_columns):
-        error_msg += f"""Mandatory columns missing in the {name} data.
-Please provide the following columns: {mandatory_columns}
-"""
-    if error_msg:
-        return error_msg
-    return ""
-
-
-def _process_data(
-    template_df: pd.DataFrame, new_campaign_df: pd.DataFrame
-) -> GoogleSheetValues:
-    final_df = pd.DataFrame(columns=template_df.columns)
-    for _, template_row in template_df.iterrows():
-        for _, new_campaign_row in new_campaign_df.iterrows():
-            campaign = f"{new_campaign_row['Country']} - {new_campaign_row['Station From']} - {new_campaign_row['Station To']}"
-            for ad_group in [
-                f"{new_campaign_row['Station From']} - {new_campaign_row['Station To']}",
-                f"{new_campaign_row['Station To']} - {new_campaign_row['Station From']}",
-            ]:
-                new_row = template_row.copy()
-                new_row["Campaign"] = campaign
-                new_row["Ad Group"] = ad_group
-                final_df = pd.concat(
-                    [final_df, pd.DataFrame([new_row])], ignore_index=True
-                )
-
-    values = [final_df.columns.tolist(), *final_df.values.tolist()]
-    return GoogleSheetValues(values=values)
-
-
 @app.post(
     "/process-data",
     description="Process data to generate new ads or keywords based on the template",
@@ -383,7 +347,7 @@ async def process_data(
             status_code=status.HTTP_400_BAD_REQUEST, detail=validation_error_msg
         )
 
-    return _process_data(template_df, new_campaign_df)
+    return process_data_f(template_df, new_campaign_df)
 
 
 @app.post(
